move test file and npa hierarchy to state_opt
