+++ toqito/helper/__init__.py
@@ -3,5 +3,5 @@
 from toqito.helper.expr_as_np_array import expr_as_np_array
 from toqito.helper.np_array_as_expr import np_array_as_expr
 from toqito.helper.update_odometer import update_odometer
-from toqito.helper.npa_hierarchy import npa_constraints
+from toqito.state_opt.npa_hierarchy import npa_constraints
 from toqito.helper.channel_dim import channel_dim
