+++ toqito/perms/tests/test_vec.py
